% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_doc_functions.R
\name{is_text_doc}
\alias{is_text_doc}
\title{Detect whether a document contains text elements}
\usage{
is_text_doc(doc_path)
}
\arguments{
\item{doc_path}{(character) Path and name of the input file. This must be a
PDF-format document.}
}
\value{
\code{TRUE} is the document is made up of scanned images rather than
  standard text elements; \code{FALSE} otherwise.
}
\description{
This function can be used to classify documents into those with standard PDF
text elements versus those with no such elements (typically documents
containing scanned images).
}
