% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_doc_functions.R
\name{read_iap_sections}
\alias{read_iap_sections}
\title{Parse the text of an Incident Action Plan document into sections}
\usage{
read_iap_sections(
  doc_path,
  sections = NULL,
  doctype = c("simple", "scanned", "all"),
  dpi = 300
)
}
\arguments{
\item{doc_path}{(character) Path and name of the IAP document (or a similarly
structured report). This must be a PDF-format document.}

\item{sections}{(character) Either \code{NULL} for a pre-defined set of IAP
section headers (default), or one or more names of sections to return. The
default section names are defined in the character vector
\code{'CERMBscraper::IAP_SECTION_NAMES'} which is included with the package.}

\item{doctype}{(character) Type of documents to process. Options are:
  \code{'simple'} (default) to only read standard documents with PDF text
  elements; \code{'scanned'} to only read documents containing images of
  scanned hard copies; or \code{'all'} to attempt to read all documents.

Set to \code{FALSE} (default) to skip
  documents containing scanned images; or \code{TRUE} to attempt to retrieve
  text using OCR. Scanned image processing often fails, and relies on each
  page of the document corresponding to a single page of the hard copy.}

\item{dpi}{(integer) Resolution to use when processing scanned image
documents (default 300).}
}
\value{
A data frame with columns: section, start_pos, end_pos, text;
  or \code{NULL} if the document does not contain any recognized sections.
}
\description{
This function has been specifically written to deal with IAP documents but
might work with other report formats as well if one or more relevant section
names can be provided via the \code{sections} argument.
}
