% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_doc_functions.R
\name{read_intel_log}
\alias{read_intel_log}
\title{Read a YAML formatted intel log file as a data frame}
\usage{
read_intel_log(log_path)
}
\arguments{
\item{log_path}{(character) Path and name of the intel log file.}
}
\value{
A data frame with a column for each of seven attributes plus an extra
  column with integer values for abbreviated grid references.
}
\description{
Rural Fire Service intel log files contain one or more situation reports
("sit reps") describing the status of a given fire and actions being
undertaken to control it at a given time. The sequence of report information
is presented in YAML format. This function reads a log file and converts the
information to tabular format (a data frame) with a column for each of seven
attributes: Subject (usually 'Sit Rep'), Priority, Reliability, DateTime,
Information (narrative text), Author and Role. In addition, the function
extracts grid references from the Information attribute and returns them as
an extra column, GridRef. Grid reference values are 6-digit integers, following
the RFS convention of reducing a full grid reference (6-digit easting and
7-digit northing) by taking digits 2-4 of the easting and digits 3-5 of the
northing, to give a location with 100m precision.
}
\examples{
\dontrun{
dat <- read_intel_log("1234567890_IntelLog.yaml")
}

}
