% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IAP_SECTION_NAMES}
\alias{IAP_SECTION_NAMES}
\title{Standard section names for Incident Action Plans}
\format{
An object of class \code{character} of length 6.
}
\usage{
IAP_SECTION_NAMES
}
\description{
A named character vector of string patterns to identify the most commonly
encountered section names. Names are lower case, single word section
identifiers.
}
\keyword{datasets}
